Now a package
With the intention of needing only the shell (or little else), the use of characters, items, spells and other features are to be implemented, requiring a better organization.

windows now support colors as well by using the colorama library. However, the default input function doesn't supports it. Therefore, a modification of the file was needed to be made and for that added into the package and this version is to be used.